FrontEnd/requirements.txt
Flask==2.2.5
flask-socketio==5.3.2
polygon-api-client==1.12.4
eventlet==0.33.0
gunicorn==20.1.0
black==23.1.0
